schedule: add support for weekend (Saturday) courses
The fall 2020 semester brought a depressing addition of a class,
BIOL-4720: Molecular Biology Lab, which has a section that
meets on a Saturday.

This change adds support for QuACS being able to support
courses with Saturday meeting times in both the section view
and calendar / schedule overview.

Per request, the section cards will only display weekend days
iff. a course has a timeslot on a weekend.