courses: add support for 'closed' indicators
Resolves #271

This commit enables support for scraping and displaying course closed indicators (defined as 'NR' - Not available for registration  or 'C' - closed on the sis view).

For the QuACS UI, a course is defined as closed when the following conditions are met:
1. All sections have 'closed':true set.
2. The course is not full.

This avoids redundant labeling since for all intents-and-purposes a full course is a closed course.
For sections, the logic is similar except the individual section attribute and size is checked.