sis-scraper: ensure depts not in schools.json are still displayed
Occassionally RPI adds a department or changes the code for an
existing department, such as when STSS and STSH where merged
together to become STSO.

This ensures that even if the schools.json is incorrect, due to
catalog desyncs, that the courses will still be displayed under
an "uncategorized" category. This reduces the urgency of manual
overrides.