schedule: remove the 'closed' attribute, re-introduce XL logic
This change scrapes the `xl_rem` (number of total cross-listed seats
remaining for each course) and modifies the frontend to support
handling full cross-listed courses.

Cross listed courses have a total amount of seats available across
all the sections. Once this is exhausted, you cannot register even
if the section has seats left. This gives people (including myself
at times) a lot of false hope that they can actually register for a course.

A few months back, we started scraping the 'closed' attributes on SIS.
However, this no longer works with the new scraper, so we have to
mimic this functionality.