sis-scraper: adjust behavior in get_semesters_to_scrape
This resolves an issue noticed where fall 2021 stopped
being scraped because we're now in December and Winter Enrichment starts in December.
This also would've been observed in spring -> arch transitions as the spring ends
in May and Arch starts in May.

To address this, if our starting month is in RPI_SEMESTER_MONTH_OFFSETS,
just rollback to the previous semester.