sis-scraper: avoid infinite loop if run in January
This is an oversight from the previous commit which fixed
back-to-back semester scraping. If we are in January,
the previous code would 'roll back' a month to 0
and then proceed in an infinite loop going negative
until interrupted or killed by the 6 hour timeout
on Github Actions.

To fix this, simply set the starting month to the
maximum in the offsets set if we hit 0.