Show restrictions on section cards
Notes sections that require you to be in a specific major/degree (e.g. almost all of GSAS), to *not* be in a specific major/degree (e.g. ARTS-2230), as well as sections that require you to be in a certain class (e.g. freshman-specific IHSS) and those that require graduate standing.
Also puts cross-listings and co-requisites in a <ul> which looks much nicer.