Show courses whose credits vary by section properly
RCOS currently shows as "1 credit" despite the fact that it can be taken for 1-4 credits. This is because the old logic just picked the first section and took the credit information from that. This change accounts for sections having different credit numbers by checking all of them and finding the minimum and the maximum credits this way.